requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
qdrant-client==1.8.0
llama-index-core==0.10.20
llama-index-vector-stores-qdrant==0.1.4
llama-index-embeddings-huggingface==0.1.4
pandas==2.2.1
pyarrow==15.0.2
fastparquet==2024.2.0
boto3==1.34.69
dspy-ai
litellm
sse-starlette